+++ stmt.go
@@ -269,7 +269,11 @@ func (r *ResultResultSet) init() *ResultResultSet {
 	if r.columnSourceTypes == nil {
 		r.columnSourceTypes = make(map[string]string)
 	}
+
+	// save number of records
 	r.count = len(r.stmtOutput.Items)
+
+	// pre-calculate column types
 	colMap := make(map[string]bool)
 	for _, item := range r.stmtOutput.Items {
 		for col, av := range item {
@@ -282,6 +286,8 @@ func (r *ResultResultSet) init() *ResultResultSet {
 			}
 		}
 	}
+
+	// save column names, sorted
 	r.columnList = make([]string, 0, len(colMap))
 	for col := range colMap {
 		r.columnList = append(r.columnList, col)
@@ -302,6 +308,8 @@ func (r *ResultResultSet) ColumnTypeScanType(index int) reflect.Type {
 }
 
 // ColumnTypeDatabaseTypeName implements driver.RowsColumnTypeDatabaseTypeName/ColumnTypeDatabaseTypeName
+//
+// @since v0.3.0 ColumnTypeDatabaseTypeName returns DynamoDB's native data types (e.g. B, N, S, SS, NS, BS, BOOL, L, M, NULL).
 func (r *ResultResultSet) ColumnTypeDatabaseTypeName(index int) string {
 	return r.columnSourceTypes[r.columnList[index]]
 }